rewires session description.
