--- addon-mirage-support/factories/session-description.js
@@ -1,5 +1,0 @@
-import { Factory } from 'ember-cli-mirage';
-
-export default Factory.extend({
-  description: (i) => `session description ${i}`,
-});
