+++ addon/components/print-course.hbs
@@ -262,7 +262,7 @@
             {{t "general.description"}}:
           </label>
           <div>
-            {{session.sessionDescription.textDescription}}
+            {{session.textDescription}}
           </div>
         </div>
       </div>
