--- app/models/session-description.js
@@ -1,1 +1,0 @@
-export { default } from 'ilios-common/models/session-description';
